+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Remove acted candidates from discovery list after approve/deny
 ---
 
 ## [0.8.0] - 2026-03-04